PROGRESS NOTE @backup=20190404113726
Lots of good work done on FileMakerUNIT for Unit-Testing FileMaker Logic. I've Copied JSON_Fun.fmp12 to STEAM_Fun.fmp12 and built a good hunch of FileMakerUNIT. Now that FileMakerUNIT is getting going, I think I will abandon the version in STEAM_Fun.fmp12 and I've copied it as FileMakerUNIT.fm12. That is what the backup is. I almost stripped out all of the JSON_Fun stuff, but realized there is no harm in taking a backup before that is done.