PROGRESS NOTE @backup=20190425133236
I accidentally closed Steam.fmp12, so good opportunity for a backup. LOTS of progress with the DailyTasks Report
Incidentally, I have begun the migration of xhtml_ Functions(), to vhtml_ Functions. That is, XHTML is not in vogue any more, but HTML5 is -- get it? 'v' as in Roman Number 'V' for 5 as in HTML5.