PUNCHOUT @backup=2019081411553
All (current) JSON files are loaded upon each login. The folder structure to hold these files is confirmed or created just prior. The (pretty big, and unlikely to change) geo-json of Colleges and Universities is scheduled to be loaded upon the first logon after the first of the month. I can switch from DEV to LIVE on a moment’s notice — but I’m not sure that is necessary(?). Only thing is that I have to compose a parallel block of code that I have written for macOS and tweak it for Windows.